+++ environment/environment.go
@@ -112,7 +112,7 @@ func InitEnvironment(configFile ConfigFile) (string, string, string, string, str
 	// initialize parent cargoport dirs on system
 	var err error
 
-	// Create /var/cargoport/ directories on local machine
+	// create /var/cargoport/ directories on local machine
 	cargoportBase := strings.TrimSuffix(configFile.DefaultCargoportDir, "/")
 	cargoportLocal := filepath.Join(cargoportBase, "/local")
 	cargoportRemote := filepath.Join(cargoportBase, "/remote")
@@ -122,18 +122,27 @@ func InitEnvironment(configFile ConfigFile) (string, string, string, string, str
 	if err = os.MkdirAll(cargoportBase, 0755); err != nil {
 		log.Fatalf("ERROR <environment>: Error creating directory %s: %v", cargoportLocal, err)
 	}
+
+	fmt.Printf("Root directory initialized at: %s\n", cargoportBase)
+
 	// create /$CARGOPORT/local
 	if err = os.MkdirAll(cargoportLocal, 0755); err != nil {
 		log.Fatalf("ERROR <environment>: Error creating directory %s: %v", cargoportLocal, err)
 	}
+	fmt.Printf("Local backup directory: %s\n", cargoportLocal)
+
 	// create /$CARGOPORT/remote
 	if err = os.MkdirAll(cargoportRemote, 0755); err != nil {
 		log.Fatalf("ERROR <environment>: Error creating directory %s: %v", cargoportRemote, err)
 	}
+	fmt.Printf("Remote backup directory: %s\n", cargoportRemote)
+
 	// create /$CARGOPORT/keys cargoportKeys
 	if err = os.MkdirAll(cargoportKeys, 0755); err != nil {
 		log.Fatalf("ERROR <environment>: Error creating directory %s: %v", cargoportKeys, err)
 	}
+	fmt.Printf("Keytool storage: %s\n", cargoportKeys)
+
 	// set 777 on /var/cargoport/remote for all users to access
 	err = sysutil.RunCommand("chmod", "-R", "777", cargoportRemote)
 	if err != nil {
@@ -142,6 +151,7 @@ func InitEnvironment(configFile ConfigFile) (string, string, string, string, str
 
 	// initialize logging
 	logFilePath := initLogging(cargoportBase)
+	fmt.Printf("Log file initialized at: %s\n", logFilePath)
 
 	return cargoportBase, cargoportLocal, cargoportRemote, logFilePath, cargoportKeys
 }
@@ -151,7 +161,7 @@ func initLogging(cargoportBase string) (logFilePath string) {
 	logFilePath = filepath.Join(cargoportBase, "cargoport-main.log")
 	logFile, err := os.OpenFile(logFilePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
-		fmt.Printf("ERROR: Failed to initialize logging: %v\n", err)
+		fmt.Printf("ERROR: Failed to initialize logging: %v", err)
 		os.Exit(1)
 	}
 	log.SetOutput(logFile)
@@ -161,19 +171,27 @@ func initLogging(cargoportBase string) (logFilePath string) {
 
 // guided setup tool for initial init
 func SetupTool() {
-	fmt.Println("---- Cargoport Setup Wizard -----")
-	fmt.Println("Thanks for trying this out!")
-	fmt.Println(" ")
+	fmt.Println("  |---- Cargoport Setup Wizard -----|")
+	fmt.Println("  |-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-|")
+	fmt.Println("      Thanks for trying this out!")
+	fmt.Println("                                 ")
+
+	// validate that current UID=0 during setuptool process
+	if os.Geteuid() != 0 {
+		log.Fatalf("ERROR <environment>: The -setup command must be run as root (e.g. with sudo).")
+	}
 
 	// prompt for root directory
 	var rootDir string
-	fmt.Println("Please specify the root directory for Cargoport's data & backupfile stores")
-	fmt.Println("/var/cargoport/ works in most cases, keep the trailing `/` here:")
-	fmt.Print("Root directory (default: /var/cargoport/): ")
+	fmt.Println("Please specify the root directory for Cargoport's data & backup storage")
+	fmt.Println("Leave blank for /var/cargoport, which works in most cases")
+	fmt.Println(" ")
+	fmt.Print("Root directory (default: /var/cargoport): ")
 	fmt.Scanln(&rootDir)
 	if rootDir == "" {
 		rootDir = "/var/cargoport/"
 	}
+	fmt.Println("------")
 	fmt.Println(" ")
 
 	// ensure that passed directory name ends in cargoport, otherwise join cargoport onto it
@@ -181,7 +199,7 @@ func SetupTool() {
 	if !strings.HasSuffix(rootDir, "cargoport") {
 		rootDir = filepath.Join(rootDir, "cargoport")
 	}
-	fmt.Printf("Using root directory: %s\n", rootDir)
+	fmt.Printf("Using root dir: %s\n", rootDir)
 	fmt.Println(" ")
 
 	// walk through temp configfile for setup & init
@@ -193,18 +211,10 @@ func SetupTool() {
 		RemoteOutputDir:     filepath.Join(rootDir, "remote/"),
 	}
 
-	// detect if setup already exists
-	//
-
 	// init env and determine directories & logfile
-	cargoportBase, cargoportLocal, cargoportRemote, logFilePath, cargoportKeys := InitEnvironment(configFile)
+	cargoportBase, _, _, _, cargoportKeys := InitEnvironment(configFile)
 
-	// print new dir and logfile information
-	fmt.Printf("Root directory initialized at: %s\n", cargoportBase)
-	fmt.Printf("Local backup directory: %s\n", cargoportLocal)
-	fmt.Printf("Remote backup directory: %s\n", cargoportRemote)
-	fmt.Printf("Log file initialized at: %s\n", logFilePath)
-	fmt.Printf("Keytool storage: %s\n", cargoportKeys)
+	fmt.Println("------")
 	fmt.Println(" ")
 
 	// check for existing config.yml
@@ -212,21 +222,33 @@ func SetupTool() {
 
 	// if DNE then prompt to create default config
 	if _, err := os.Stat(configFilePath); os.IsNotExist(err) {
-		var createConfig string
-		fmt.Printf("No config.yml found in %s. Would you like to create one? (y/n): ", cargoportBase)
-		fmt.Scanln(&createConfig) // scan for input
-		// if yes then invoke createDefaultConfig
-		if strings.ToLower(createConfig) == "y" {
-			err := createDefaultConfig(configFilePath, rootDir)
-			if err != nil {
-				log.Fatalf("ERROR: Failed to create config.yml: %v\n", err)
+		for {
+			fmt.Printf("No config.yml found in %s. Would you like to create one? (y/n): ", cargoportBase)
+			var createConfig string
+			// scan for input
+			fmt.Scanln(&createConfig)
+
+			// switch to catch input choices
+			switch strings.ToLower(createConfig) {
+			case "y":
+				err := createDefaultConfig(configFilePath, rootDir)
+				if err != nil {
+					log.Fatalf("ERROR: Failed to create config.yml %v", err)
+				}
+				log.Println("INFO <environment>: Default config.yml created at %s", configFilePath)
+				fmt.Println("Default config.yml created at %s", configFilePath)
+				break
+
+			case "n":
+				log.Println("WARN <environment>: Skipping automatic configfile creation, please ensure you manually create a config.yml file!")
+				break
+
+			default:
+				fmt.Println("Invalid input, please try again")
+				continue // loop again to prompt for input
 			}
-			fmt.Printf("Default config.yml created at %s.\n", configFilePath)
-		} else {
-			fmt.Println("Skipping config.yml creation.")
+			break
 		}
-	} else {
-		fmt.Println("Detected existing config.yml in parent directory")
 	}
 	fmt.Println(" ")
 
@@ -238,8 +260,9 @@ func SetupTool() {
 
 	// save true config at /etc/ reference
 	if err := saveTrueConfigReference(configFilePath); err != nil {
-		log.Fatalf("ERROR: Failed to save true config reference: %v\n", err)
+		log.Fatalf("ERROR: Failed to save true config reference: %v", err)
 	}
+	fmt.Println("------")
 	fmt.Println(" ")
 
 	fmt.Println("Environment setup completed successfully.")